Relocate BenchmarkContext into simulation/; free contexts/; drop dead get_subcontext code
The `contexts/` package held tabrepo-era machinery for building an
EvaluationRepository / ZeroshotSimulatorContext from downloaded zeroshot data
(`BenchmarkContext`, `construct_context`, artificial fixtures) — distinct from
the `*ArenaContext` evaluation family. Its dependencies are overwhelmingly
`simulation` (+ repository/predictions), and its only consumers are a lazy
import in `repository/` and a few repository/predictions tests, so it belongs
under `simulation/`. Moving it frees the top-level `contexts/` name for the
Arena context family (follow-up PR).

- `contexts/context.py` -> `simulation/benchmark_context.py`
- `contexts/context_artificial.py` -> `simulation/context_artificial.py`
- `contexts/utils.py`: folded its sole function `load_zeroshot_input` into
  `benchmark_context.py` (its only caller) and deleted the file.
- `repository/evaluation_repository_collection.py` now imports
  `prune_zeroshot_gt` straight from `simulation/dense_utils` (it was only
  re-exported through `contexts.utils`).
- `repository/evaluation_repository.py` lazy imports updated to the new path
  (kept lazy — it breaks a module-level cycle).
- Tests updated to import `context_artificial` from `simulation/`.

Also remove dead `get_subcontext` logic: `sim_runner.run_zs_sim_end_to_end`
was unused and its main path called `get_subcontext`, which is not defined
anywhere (the subcontext machinery is gone) — so the function raised on use.
Dropped the function and its now-orphaned `catchtime` import.
