--- packages/tabarena/src/tabarena/contexts/__init__.py
@@ -1,3 +1,0 @@
-from __future__ import annotations
-
-from .context import BenchmarkContext
