+++ tests/tabarena/repository/test_repository_collection.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 from tabarena import EvaluationRepositoryCollection
-from tabarena.contexts.context_artificial import load_repo_artificial
+from tabarena.simulation.context_artificial import load_repo_artificial
 
 from .test_repository import verify_equivalent_repository
